+++ tests/testdata/suppressions/dangling.py
@@ -1,0 +1,3 @@
+# <block name="source" affects=":gone_a,:gone_b">
+VALUE = 1
+# </block>